[MOD] patched error edit field admin in update method AdminUser + protected all admin pages from guest

+++ FreeAds/routes/web.php
@@ -33,10 +33,10 @@ Route::get('/admin/users', [App\Http\Controllers\Admin\AdminUserController::clas
 ->middleware('is_admin')
 ->name('admin.users.index')
 ;
-Route::post('/admin/users', [App\Http\Controllers\Admin\AdminUserController::class, 'store'])->name('admin.users.index');
-Route::get('/admin/users/create', [App\Http\Controllers\Admin\AdminUserController::class, 'create'])->name('admin.users.create');
-Route::get('/admin/users/edit/{id}', [App\Http\Controllers\Admin\AdminUserController::class, 'edit'])->name('admin.users.edit');
-Route::post('/admin/users/update/{id}', [App\Http\Controllers\Admin\AdminUserController::class, 'update'])->name('admin.users.update');
-Route::delete('/admin/users/delete/{id}', [App\Http\Controllers\Admin\AdminUserController::class, 'destroy'])->name('admin.users.delete');
+Route::post('/admin/users', [App\Http\Controllers\Admin\AdminUserController::class, 'store'])->middleware('is_admin')->name('admin.users.index');
+Route::get('/admin/users/create', [App\Http\Controllers\Admin\AdminUserController::class, 'create'])->middleware('is_admin')->name('admin.users.create');
+Route::get('/admin/users/edit/{id}', [App\Http\Controllers\Admin\AdminUserController::class, 'edit'])->middleware('is_admin')->name('admin.users.edit');
+Route::post('/admin/users/update/{id}', [App\Http\Controllers\Admin\AdminUserController::class, 'update'])->middleware('is_admin')->name('admin.users.update');
+Route::delete('/admin/users/delete/{id}', [App\Http\Controllers\Admin\AdminUserController::class, 'destroy'])->middleware('is_admin')->name('admin.users.delete');
